backend/pyproject.toml
[tool.poetry]
name = "agent-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <first.last@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
langchain = "^0.3.6"
langchain-openai = "^0.2.5"
langgraph = "^0.2.42"
pydantic = "^2.9.2"
cdp-sdk = "^0.10.3"
python-dotenv = "^1.0.1"
cdp-langchain = "^0.0.4"
cdp-agentkit-core = "^0.0.3"
flask-cors = "^5.0.0"
flask-sock = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
